Fixes for BlockSizeDeduction. Iffy, people should definitely look at them
